Install the new exchange in the Docker image
The significant change is in the Dockerfile, where we install poetry so we can install the new exchange (and possibly future ones as well). We try to keep the number of layers relatively small to make the image easier to download/less bloated.

There's also some indirection via Docker Compose Secrets here for local testing/debugging. Essentially, we're just trying to run the environment variable from the machine down to the container, which requires secrets for reasons. Unfortunately, secrets can only write to files inside the container, so we have to extract the contents of that file back into the environment. That extraction is done in `startup.sh`.